+++ src/iop/__init__.py
@@ -10,6 +10,7 @@ from iop._message import (
     _PydanticPickleMessage,
 )
 from iop._outbound_adapter import _OutboundAdapter
+from iop._polling_business_service import _PollingBusinessServiceMixin
 from iop._persistent_message import Field as Field
 from iop._persistent_message import Model as Model
 from iop._persistent_message import _PersistentMessage
@@ -31,6 +32,10 @@ class OutboundAdapter(_OutboundAdapter):
 
 
 class BusinessService(_BusinessService):
+    pass
+
+
+class PollingBusinessService(_PollingBusinessServiceMixin, BusinessService):
     pass